Moved ChatuserContext to AppContext
